+++ lib/src/styles/skadi_responsive.dart
@@ -38,11 +38,13 @@ class SkadiResponsive {
   static Size? _size;
   static BuildContext? context;
 
-  static SkadiResponsiveBreakpoint _breakPoint = SkadiResponsiveBreakpoint.defaultValue();
+  static SkadiResponsiveBreakpoint _breakPoint =
+      SkadiResponsiveBreakpoint.defaultValue();
 
   static double get screenWidth {
     if (_size == null) {
-      throw FlutterError('Please initialize SkadiResponsiveBuilder in MaterialApp builder');
+      throw FlutterError(
+          'Please initialize SkadiResponsiveBuilder in MaterialApp builder');
     }
     return _size!.width;
   }
@@ -58,7 +60,8 @@ class SkadiResponsive {
     _breakPoint = breakPoint;
   }
 
-  static SkadiResponsiveBreakpointName _getBreakpointName([BuildContext? context]) {
+  static SkadiResponsiveBreakpointName _getBreakpointName(
+      [BuildContext? context]) {
     double modifiedWidth = 0.0;
     if (context != null) {
       modifiedWidth = MediaQuery.of(context).size.width;
@@ -77,7 +80,8 @@ class SkadiResponsive {
 
   static bool get isDesktop => screenWidth >= _breakPoint.desktop;
   static bool get isTablet => !isDesktop && screenWidth >= _breakPoint.tablet;
-  static bool get isMobile => screenWidth > _breakPoint.mobileSmall && screenWidth < _breakPoint.tablet;
+  static bool get isMobile =>
+      screenWidth > _breakPoint.mobileSmall && screenWidth < _breakPoint.tablet;
   static bool get isMobileSmall => screenWidth <= _breakPoint.mobileSmall;
 
   ///Build a widget base on device screen size