+++ lib/src/utilities/logger.dart
@@ -17,12 +17,14 @@ class SkadiLogSetting {
 
 void httpLog([dynamic log, dynamic additional = "", bool? logInReleaseMode]) {
   var setting = SkadiProvider.loggerSetting;
-  if (kDebugMode || (logInReleaseMode ?? setting.http)) debugPrint("Http Log: $log $additional");
+  if (kDebugMode || (logInReleaseMode ?? setting.http))
+    debugPrint("Http Log: $log $additional");
 }
 
 void infoLog([dynamic log, dynamic additional = "", bool? logInReleaseMode]) {
   var setting = SkadiProvider.loggerSetting;
-  if (kDebugMode || (logInReleaseMode ?? setting.http)) debugPrint("Info Log: $log $additional");
+  if (kDebugMode || (logInReleaseMode ?? setting.http))
+    debugPrint("Info Log: $log $additional");
 }
 
 void errorLog([dynamic log, dynamic additional = ""]) {