+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## [0.1.3] - 16 September 2022
+## [0.2.0] - 28 September 2022
 
 - Improve README
+- Add setting for `Logger`
 - Add `SkadiBadge`
+- breaking change on `SkadiAsyncButton` to follow ElevatedButton parameter name
 
 ## [0.1.2] - 13 September 2022